+++ src/components/forms/CampaignForm.tsx
@@ -1,8 +1,7 @@
-import React from "react";
+import React, { CSSProperties, useState } from "react";
 import {
   chakra,
   Box,
-  Flex,
   useColorModeValue,
   SimpleGrid,
   GridItem,
@@ -17,19 +16,75 @@ import {
   FormHelperText,
   Textarea,
   Avatar,
-  Icon,
-  Button,
-  VisuallyHidden,
   Select,
-  Checkbox,
-  RadioGroup,
-  Radio,
 } from "@chakra-ui/react";
-import { FaUser } from "react-icons/fa";
+import { EmailIcon } from '@chakra-ui/icons'
+import { useDropzone } from 'react-dropzone'
+import LoomRecordButton from 'src/components/LoomRecordButton'
+import { TextileInstance } from '../../services/textile/textile'
 
-export default function CampaignForm() {
+const imgUploadContainer: CSSProperties = {
+    marginLeft: '0.25rem',
+  }
+  
+  const dropzoneContainer: CSSProperties = {
+    width: '450px',
+    height: '107px',
+    display: 'flex',
+    justifyItems: 'start',
+    alignItems: 'start',
+    backgroundImage: `url("data:image/svg+xml,%3csvg width='100%25' height='100%25' xmlns='http://www.w3.org/2000/svg'%3e%3crect width='100%25' height='100%25' fill='none' stroke='grey' stroke-width='3' stroke-dasharray='2%2c 6' stroke-dashoffset='0' stroke-linecap='square'/%3e%3c/svg%3e")`,
+  }
+  
+  const dropzone: CSSProperties = {
+    marginTop: '2px',
+    padding: 'auto',
+    display: 'flex',
+    justifyContent: 'start',
+    alignContent: 'start',
+    alignItems: 'center',
+    justifyItems: 'center',
+    width: '446px',
+    height: '103px',
+    marginLeft: '2px',
+    paddingLeft: '1rem',
+  }
+  
+  const thumbsContainer: CSSProperties = {
+    width: '450px',
+    marginBottom: '0.5rem',
+    display: 'flex',
+    alignItems: 'center',
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+  }
+  
+  const thumbLabel: CSSProperties = {
+    marginLeft: '0.75rem',
+    marginRight: '1.25rem',
+    display: 'flex',
+    alignItems: 'center',
+    justifyItems: 'center',
+  }
+
+export default function CampaignForm(props) {
+    const [videoUrl, setVideoUrl] = useState('')
+    const [files, setFiles] = useState([])
+
+    const { getRootProps, getInputProps } = useDropzone({
+        onDrop: (acceptedFiles) => {
+          setFiles(
+            acceptedFiles.map((file) =>
+              Object.assign(file, {
+                preview: URL.createObjectURL(file),
+              })
+            )
+          )
+        },
+      })
+
   return (
-    <Box bg={useColorModeValue("gray.100", "inherit")} p={10}>
+    <Box p={10}>
       <Box>
         <SimpleGrid
           display={{ base: "initial", md: "grid" }}
@@ -67,429 +122,544 @@ export default function CampaignForm() {
                 p={{ sm: 6 }}
               >
                 <SimpleGrid columns={3} spacing={6}>
-                  <FormControl as={GridItem} colSpan={[3, 2]}>
-                    <FormLabel
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      Website
-                    </FormLabel>
-                    <InputGroup size="sm">
-                      <InputLeftAddon
-                        children="http://"
-                        bg={useColorModeValue("gray.500", "gray.800")}
-                        color={useColorModeValue("gray.500", "gay.50")}
-                        rounded="md"
-                      />
+                    <FormControl as={GridItem} colSpan={[3, 2]}>
+                      <FormLabel
+                        fontSize="sm"
+                        fontWeight="md"
+                        color={useColorModeValue('gray.700', 'gray.50')}
+                        htmlFor="brandName"
+                      >
+                        Brand Name
+                      </FormLabel>
                       <Input
+                        value={props.brandName}
+                        onChange={props.handleChange}
                         type="text"
-                        placeholder="www.example.com"
+                        placeholder="Company Name"
+                        id="brandName"
+                        name="brandName"
                         focusBorderColor="brand.400"
                         rounded="md"
-                        color="gray.700"
+                        color={useColorModeValue('gray.700', 'gray.50')}
                       />
-                    </InputGroup>
-                  </FormControl>
-                </SimpleGrid>
-
-                  <FormControl id="email" mt={1}>
-                    <FormLabel
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      About
-                    </FormLabel>
-                    <Textarea
-                      placeholder="you@example.com"
-                      mt={1}
-                      rows={3}
-                      shadow="sm"
-                      focusBorderColor="brand.400"
-                      fontSize={{ sm: "sm" }}
-                    />
-                    <FormHelperText>
-                      Brief description for your profile. URLs are hyperlinked.
-                    </FormHelperText>
-                  </FormControl>
-
-                <FormControl>
-                  <FormLabel
-                    fontSize="sm"
-                    fontWeight="md"
-                    color={useColorModeValue("gray.700", "gray.50")}
-                  >
-                    Photo
-                  </FormLabel>
-                  <Flex alignItems="center" mt={1}>
-                    <Avatar
-                      boxSize={12}
-                      bg={useColorModeValue("gray.100", "gray.800")}
-                      icon={
-                        <Icon
-                          as={FaUser}
-                          boxSize={9}
-                          mt={3}
-                          rounded="full"
-                          color={useColorModeValue("gray.300", "gray.700")}
-                        />
-                      }
-                    />
-                    <Button
-                    colorScheme="blue"
-                      ml={5}
-                      variant="outline"
-                      size="sm"
-                      fontWeight="medium"
-                      _focus={{ shadow: "none" }}
-                    >
-                      Change
-                    </Button>
-                  </Flex>
-                </FormControl>
-
-                <FormControl>
-                  <FormLabel
-                    fontSize="sm"
-                    fontWeight="md"
-                    color={useColorModeValue("gray.700", "gray.50")}
-                  >
-                    Cover photo
-                  </FormLabel>
-                  <Flex
-                    mt={1}
-                    justify="center"
-                    px={6}
-                    pt={5}
-                    pb={6}
-                    borderWidth={2}
-                    borderColor={useColorModeValue("gray.300", "gray.500")}
-                    borderStyle="dashed"
-                    rounded="md"
-                  >
-                    <Stack spacing={1} textAlign="center">
-                      <Icon
-                        mx="auto"
-                        boxSize={12}
-                        color={useColorModeValue("gray.400", "gray.500")}
-                        stroke="currentColor"
-                        fill="none"
-                        viewBox="0 0 48 48"
-                        aria-hidden="true"
-                      >
-                        <path
-                          d="M28 8H12a4 4 0 00-4 4v20m32-12v8m0 0v8a4 4 0 01-4 4H12a4 4 0 01-4-4v-4m32-4l-3.172-3.172a4 4 0 00-5.656 0L28 28M8 32l9.172-9.172a4 4 0 015.656 0L28 28m0 0l4 4m4-24h8m-4-4v8m-12 4h.02"
-                          strokeWidth="2"
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                        />
-                      </Icon>
-                      <Flex
+                    </FormControl>
+                  </SimpleGrid>
+                  <SimpleGrid columns={3} spacing={6}>
+                    <FormControl as={GridItem} colSpan={[3, 2]}>
+                      <FormLabel
                         fontSize="sm"
-                        color={useColorModeValue("gray.600", "gray.400")}
-                        alignItems="baseline"
+                        fontWeight="md"
+                        color={useColorModeValue('gray.700', 'gray.50')}
                       >
-                        <chakra.label
-                          htmlFor="file-upload"
-                          cursor="pointer"
+                        Brand Website
+                      </FormLabel>
+                      <InputGroup size="sm">
+                        <InputLeftAddon
+                          children="https://"
+                          bg={useColorModeValue('gray.50', 'gray.800')}
+                          color={useColorModeValue('gray.500', 'gay.50')}
                           rounded="md"
-                          fontSize="md"
-                          color={useColorModeValue("brand.600", "brand.200")}
-                          pos="relative"
-                          _hover={{
-                            color: useColorModeValue("brand.400", "brand.300"),
-                          }}
-                        >
-                          <span>Upload a file</span>
-                          <VisuallyHidden>
-                            <input
-                              id="file-upload"
-                              name="file-upload"
-                              type="file"
-                            />
-                          </VisuallyHidden>
-                        </chakra.label>
-                        <Text pl={1}>or drag and drop</Text>
-                      </Flex>
-                      <Text
-                        fontSize="xs"
-                        color={useColorModeValue("gray.500", "gray.50")}
+                        />
+                        <Input
+                          value={props.brandWebsite}
+                          onChange={props.handleChange}
+                          type="text"
+                          placeholder="www.example.com"
+                          id="brandWebsite"
+                          name="brandWebsite"
+                          focusBorderColor="brand.400"
+                          rounded="md"
+                          color={useColorModeValue('gray.700', 'gray.50')}
+                        />
+                      </InputGroup>
+                    </FormControl>
+                  </SimpleGrid>
+
+                  <SimpleGrid columns={3} spacing={6}>
+                    <FormControl as={GridItem} colSpan={[3, 2]}>
+                      <FormLabel
+                        fontSize="sm"
+                        fontWeight="md"
+                        color={useColorModeValue('gray.700', 'gray.50')}
                       >
-                        PNG, JPG, GIF up to 10MB
-                      </Text>
-                    </Stack>
-                  </Flex>
-                </FormControl>
-              </Stack>
-              <Box
-                px={{ base: 4, sm: 6 }}
-                py={3}
-                bg={useColorModeValue("gray.200", "gray.900")}
-                textAlign="right"
-              >
-                <Button
-                  type="submit"
-                  colorScheme="brand"
-                  _focus={{ shadow: "" }}
-                  fontWeight="md"
-                >
-                  Save
-                </Button>
-              </Box>
+                        Twitter Username
+                      </FormLabel>
+                      <InputGroup size="sm">
+                        <InputLeftAddon
+                          children="@"
+                          bg={useColorModeValue('gray.50', 'gray.800')}
+                          color={useColorModeValue('gray.500', 'gay.50')}
+                          rounded="md"
+                        />
+                        <Input
+                          value={props.twitterAccount}
+                          onChange={props.handleChange}
+                          id="twitterAccount"
+                          name="twitterAccount"
+                          placeholder="creativecrtv"
+                          focusBorderColor="brand.400"
+                          rounded="md"
+                          color={useColorModeValue('gray.700', 'gray.50')}
+                        />
+                      </InputGroup>
+                      <FormHelperText>
+                        Verify your account via Twitter (0.1 LINK).
+                      </FormHelperText>
+                    </FormControl>
+                  </SimpleGrid>
+
+                  <SimpleGrid columns={3} spacing={6}>
+                    <FormControl mt={1} as={GridItem} colSpan={[3, 2]}>
+                      <FormLabel
+                        htmlFor="campaignBrief"
+                        fontSize="sm"
+                        fontWeight="md"
+                        color={useColorModeValue('gray.700', 'gray.50')}
+                      >
+                        Campaign Brief
+                      </FormLabel>
+                      <Textarea
+                        value={props.campaignBrief}
+                        onChange={props.handleChange}
+                        id="campaignBrief"
+                        name="campaignBrief"
+                        placeholder="Additional information about your campaign"
+                        mt={1}
+                        rows={3}
+                        shadow="sm"
+                        focusBorderColor="brand.400"
+                        fontSize={{
+                          sm: 'sm',
+                        }}
+                        color={useColorModeValue('gray.700', 'gray.50')}
+                      />
+                      <FormHelperText>
+                        Brief description for your requirements. URLs are
+                        hyperlinked.
+                      </FormHelperText>
+                    </FormControl>
+                  </SimpleGrid>
+
+                  <SimpleGrid columns={3} spacing={6}>
+                    <FormControl as={GridItem} colSpan={[3, 2]}>
+                      <FormLabel
+                        htmlFor="email"
+                        fontSize="sm"
+                        fontWeight="md"
+                        color={useColorModeValue('gray.700', 'gray.50')}
+                      >
+                        Email address
+                      </FormLabel>
+                      <InputGroup size={'sm'}>
+                        <InputLeftAddon
+                          children={<EmailIcon />}
+                          bg={useColorModeValue('gray.50', 'gray.800')}
+                          color={useColorModeValue('gray.500', 'gay.50')}
+                          rounded="md"
+                        />
+                        <Input
+                          type="text"
+                          name="email"
+                          id="email"
+                          value={props.email}
+                          onChange={props.handleChange}
+                          autoComplete="email"
+                          focusBorderColor="brand.400"
+                          w="full"
+                          rounded="md"
+                          color={useColorModeValue('gray.700', 'gray.50')}
+                        />
+                      </InputGroup>
+                    </FormControl>
+                  </SimpleGrid>
+
+                  <SimpleGrid columns={3} spacing={6}>
+                    <FormControl>
+                      <FormLabel
+                        htmlFor="country"
+                        fontSize="sm"
+                        fontWeight="md"
+                        color={useColorModeValue('gray.700', 'gray.50')}
+                      >
+                        Country / Region
+                      </FormLabel>
+                      <Select
+                        value={props.country}
+                        onChange={props.handleChange}
+                        id="country"
+                        name="country"
+                        autoComplete="country"
+                        placeholder="select country"
+                        mt={1}
+                        focusBorderColor="brand.400"
+                        shadow="sm"
+                        size="sm"
+                        w="full"
+                        rounded="md"
+                        color={useColorModeValue('gray.700', 'gray.50')}
+                      >
+                        <option value="Afganistan">Afghanistan</option>
+                        <option value="Albania">Albania</option>
+                        <option value="Algeria">Algeria</option>
+                        <option value="American Samoa">American Samoa</option>
+                        <option value="Andorra">Andorra</option>
+                        <option value="Angola">Angola</option>
+                        <option value="Anguilla">Anguilla</option>
+                        <option value="Antigua &amp; Barbuda">
+                          Antigua &amp; Barbuda
+                        </option>
+                        <option value="Argentina">Argentina</option>
+                        <option value="Armenia">Armenia</option>
+                        <option value="Aruba">Aruba</option>
+                        <option value="Australia">Australia</option>
+                        <option value="Austria">Austria</option>
+                        <option value="Azerbaijan">Azerbaijan</option>
+                        <option value="Bahamas">Bahamas</option>
+                        <option value="Bahrain">Bahrain</option>
+                        <option value="Bangladesh">Bangladesh</option>
+                        <option value="Barbados">Barbados</option>
+                        <option value="Belarus">Belarus</option>
+                        <option value="Belgium">Belgium</option>
+                        <option value="Belize">Belize</option>
+                        <option value="Benin">Benin</option>
+                        <option value="Bermuda">Bermuda</option>
+                        <option value="Bhutan">Bhutan</option>
+                        <option value="Bolivia">Bolivia</option>
+                        <option value="Bonaire">Bonaire</option>
+                        <option value="Bosnia &amp; Herzegovina">
+                          Bosnia &amp; Herzegovina
+                        </option>
+                        <option value="Botswana">Botswana</option>
+                        <option value="Brazil">Brazil</option>
+                        <option value="British Indian Ocean Ter">
+                          British Indian Ocean Ter
+                        </option>
+                        <option value="Brunei">Brunei</option>
+                        <option value="Bulgaria">Bulgaria</option>
+                        <option value="Burkina Faso">Burkina Faso</option>
+                        <option value="Burundi">Burundi</option>
+                        <option value="Cambodia">Cambodia</option>
+                        <option value="Cameroon">Cameroon</option>
+                        <option value="Canada">Canada</option>
+                        <option value="Canary Islands">Canary Islands</option>
+                        <option value="Cape Verde">Cape Verde</option>
+                        <option value="Cayman Islands">Cayman Islands</option>
+                        <option value="Central African Republic">
+                          Central African Republic
+                        </option>
+                        <option value="Chad">Chad</option>
+                        <option value="Channel Islands">Channel Islands</option>
+                        <option value="Chile">Chile</option>
+                        <option value="China">China</option>
+                        <option value="Christmas Island">
+                          Christmas Island
+                        </option>
+                        <option value="Cocos Island">Cocos Island</option>
+                        <option value="Colombia">Colombia</option>
+                        <option value="Comoros">Comoros</option>
+                        <option value="Congo">Congo</option>
+                        <option value="Cook Islands">Cook Islands</option>
+                        <option value="Costa Rica">Costa Rica</option>
+                        <option value="Cote DIvoire">Cote DIvoire</option>
+                        <option value="Croatia">Croatia</option>
+                        <option value="Cuba">Cuba</option>
+                        <option value="Curaco">Curacao</option>
+                        <option value="Cyprus">Cyprus</option>
+                        <option value="Czech Republic">Czech Republic</option>
+                        <option value="Denmark">Denmark</option>
+                        <option value="Djibouti">Djibouti</option>
+                        <option value="Dominica">Dominica</option>
+                        <option value="Dominican Republic">
+                          Dominican Republic
+                        </option>
+                        <option value="East Timor">East Timor</option>
+                        <option value="Ecuador">Ecuador</option>
+                        <option value="Egypt">Egypt</option>
+                        <option value="El Salvador">El Salvador</option>
+                        <option value="Equatorial Guinea">
+                          Equatorial Guinea
+                        </option>
+                        <option value="Eritrea">Eritrea</option>
+                        <option value="Estonia">Estonia</option>
+                        <option value="Ethiopia">Ethiopia</option>
+                        <option value="Falkland Islands">
+                          Falkland Islands
+                        </option>
+                        <option value="Faroe Islands">Faroe Islands</option>
+                        <option value="Fiji">Fiji</option>
+                        <option value="Finland">Finland</option>
+                        <option value="France">France</option>
+                        <option value="French Guiana">French Guiana</option>
+                        <option value="French Polynesia">
+                          French Polynesia
+                        </option>
+                        <option value="French Southern Ter">
+                          French Southern Ter
+                        </option>
+                        <option value="Gabon">Gabon</option>
+                        <option value="Gambia">Gambia</option>
+                        <option value="Georgia">Georgia</option>
+                        <option value="Germany">Germany</option>
+                        <option value="Ghana">Ghana</option>
+                        <option value="Gibraltar">Gibraltar</option>
+                        <option value="Great Britain">Great Britain</option>
+                        <option value="Greece">Greece</option>
+                        <option value="Greenland">Greenland</option>
+                        <option value="Grenada">Grenada</option>
+                        <option value="Guadeloupe">Guadeloupe</option>
+                        <option value="Guam">Guam</option>
+                        <option value="Guatemala">Guatemala</option>
+                        <option value="Guinea">Guinea</option>
+                        <option value="Guyana">Guyana</option>
+                        <option value="Haiti">Haiti</option>
+                        <option value="Hawaii">Hawaii</option>
+                        <option value="Honduras">Honduras</option>
+                        <option value="Hong Kong">Hong Kong</option>
+                        <option value="Hungary">Hungary</option>
+                        <option value="Iceland">Iceland</option>
+                        <option value="Indonesia">Indonesia</option>
+                        <option value="India">India</option>
+                        <option value="Iran">Iran</option>
+                        <option value="Iraq">Iraq</option>
+                        <option value="Ireland">Ireland</option>
+                        <option value="Isle of Man">Isle of Man</option>
+                        <option value="Israel">Israel</option>
+                        <option value="Italy">Italy</option>
+                        <option value="Jamaica">Jamaica</option>
+                        <option value="Japan">Japan</option>
+                        <option value="Jordan">Jordan</option>
+                        <option value="Kazakhstan">Kazakhstan</option>
+                        <option value="Kenya">Kenya</option>
+                        <option value="Kiribati">Kiribati</option>
+                        <option value="Korea North">Korea North</option>
+                        <option value="Korea Sout">Korea South</option>
+                        <option value="Kuwait">Kuwait</option>
+                        <option value="Kyrgyzstan">Kyrgyzstan</option>
+                        <option value="Laos">Laos</option>
+                        <option value="Latvia">Latvia</option>
+                        <option value="Lebanon">Lebanon</option>
+                        <option value="Lesotho">Lesotho</option>
+                        <option value="Liberia">Liberia</option>
+                        <option value="Libya">Libya</option>
+                        <option value="Liechtenstein">Liechtenstein</option>
+                        <option value="Lithuania">Lithuania</option>
+                        <option value="Luxembourg">Luxembourg</option>
+                        <option value="Macau">Macau</option>
+                        <option value="Macedonia">Macedonia</option>
+                        <option value="Madagascar">Madagascar</option>
+                        <option value="Malaysia">Malaysia</option>
+                        <option value="Malawi">Malawi</option>
+                        <option value="Maldives">Maldives</option>
+                        <option value="Mali">Mali</option>
+                        <option value="Malta">Malta</option>
+                        <option value="Marshall Islands">
+                          Marshall Islands
+                        </option>
+                        <option value="Martinique">Martinique</option>
+                        <option value="Mauritania">Mauritania</option>
+                        <option value="Mauritius">Mauritius</option>
+                        <option value="Mayotte">Mayotte</option>
+                        <option value="Mexico">Mexico</option>
+                        <option value="Midway Islands">Midway Islands</option>
+                        <option value="Moldova">Moldova</option>
+                        <option value="Monaco">Monaco</option>
+                        <option value="Mongolia">Mongolia</option>
+                        <option value="Montserrat">Montserrat</option>
+                        <option value="Morocco">Morocco</option>
+                        <option value="Mozambique">Mozambique</option>
+                        <option value="Myanmar">Myanmar</option>
+                        <option value="Nambia">Nambia</option>
+                        <option value="Nauru">Nauru</option>
+                        <option value="Nepal">Nepal</option>
+                        <option value="Netherland Antilles">
+                          Netherland Antilles
+                        </option>
+                        <option value="Netherlands">
+                          Netherlands (Holland, Europe)
+                        </option>
+                        <option value="Nevis">Nevis</option>
+                        <option value="New Caledonia">New Caledonia</option>
+                        <option value="New Zealand">New Zealand</option>
+                        <option value="Nicaragua">Nicaragua</option>
+                        <option value="Niger">Niger</option>
+                        <option value="Nigeria">Nigeria</option>
+                        <option value="Niue">Niue</option>
+                        <option value="Norfolk Island">Norfolk Island</option>
+                        <option value="Norway">Norway</option>
+                        <option value="Oman">Oman</option>
+                        <option value="Pakistan">Pakistan</option>
+                        <option value="Palau Island">Palau Island</option>
+                        <option value="Palestine">Palestine</option>
+                        <option value="Panama">Panama</option>
+                        <option value="Papua New Guinea">
+                          Papua New Guinea
+                        </option>
+                        <option value="Paraguay">Paraguay</option>
+                        <option value="Peru">Peru</option>
+                        <option value="Phillipines">Philippines</option>
+                        <option value="Pitcairn Island">Pitcairn Island</option>
+                        <option value="Poland">Poland</option>
+                        <option value="Portugal">Portugal</option>
+                        <option value="Puerto Rico">Puerto Rico</option>
+                        <option value="Qatar">Qatar</option>
+                        <option value="Republic of Montenegro">
+                          Republic of Montenegro
+                        </option>
+                        <option value="Republic of Serbia">
+                          Republic of Serbia
+                        </option>
+                        <option value="Reunion">Reunion</option>
+                        <option value="Romania">Romania</option>
+                        <option value="Russia">Russia</option>
+                        <option value="Rwanda">Rwanda</option>
+                        <option value="St Barthelemy">St Barthelemy</option>
+                        <option value="St Eustatius">St Eustatius</option>
+                        <option value="St Helena">St Helena</option>
+                        <option value="St Kitts-Nevis">St Kitts-Nevis</option>
+                        <option value="St Lucia">St Lucia</option>
+                        <option value="St Maarten">St Maarten</option>
+                        <option value="St Pierre &amp; Miquelon">
+                          St Pierre &amp; Miquelon
+                        </option>
+                        <option value="St Vincent &amp; Grenadines">
+                          St Vincent &amp; Grenadines
+                        </option>
+                        <option value="Saipan">Saipan</option>
+                        <option value="Samoa">Samoa</option>
+                        <option value="Samoa American">Samoa American</option>
+                        <option value="San Marino">San Marino</option>
+                        <option value="Sao Tome &amp; Principe">
+                          Sao Tome &amp; Principe
+                        </option>
+                        <option value="Saudi Arabia">Saudi Arabia</option>
+                        <option value="Senegal">Senegal</option>
+                        <option value="Seychelles">Seychelles</option>
+                        <option value="Sierra Leone">Sierra Leone</option>
+                        <option value="Singapore">Singapore</option>
+                        <option value="Slovakia">Slovakia</option>
+                        <option value="Slovenia">Slovenia</option>
+                        <option value="Solomon Islands">Solomon Islands</option>
+                        <option value="Somalia">Somalia</option>
+                        <option value="South Africa">South Africa</option>
+                        <option value="Spain">Spain</option>
+                        <option value="Sri Lanka">Sri Lanka</option>
+                        <option value="Sudan">Sudan</option>
+                        <option value="Suriname">Suriname</option>
+                        <option value="Swaziland">Swaziland</option>
+                        <option value="Sweden">Sweden</option>
+                        <option value="Switzerland">Switzerland</option>
+                        <option value="Syria">Syria</option>
+                        <option value="Tahiti">Tahiti</option>
+                        <option value="Taiwan">Taiwan</option>
+                        <option value="Tajikistan">Tajikistan</option>
+                        <option value="Tanzania">Tanzania</option>
+                        <option value="Thailand">Thailand</option>
+                        <option value="Togo">Togo</option>
+                        <option value="Tokelau">Tokelau</option>
+                        <option value="Tonga">Tonga</option>
+                        <option value="Trinidad &amp; Tobago">
+                          Trinidad &amp; Tobago
+                        </option>
+                        <option value="Tunisia">Tunisia</option>
+                        <option value="Turkey">Turkey</option>
+                        <option value="Turkmenistan">Turkmenistan</option>
+                        <option value="Turks &amp; Caicos Is">
+                          Turks &amp; Caicos Is
+                        </option>
+                        <option value="Tuvalu">Tuvalu</option>
+                        <option value="Uganda">Uganda</option>
+                        <option value="United Kingdom">United Kingdom</option>
+                        <option value="Ukraine">Ukraine</option>
+                        <option value="United Arab Erimates">
+                          United Arab Emirates
+                        </option>
+                        <option value="United States of America">
+                          United States of America
+                        </option>
+                        <option value="Uraguay">Uruguay</option>
+                        <option value="Uzbekistan">Uzbekistan</option>
+                        <option value="Vanuatu">Vanuatu</option>
+                        <option value="Vatican City State">
+                          Vatican City State
+                        </option>
+                        <option value="Venezuela">Venezuela</option>
+                        <option value="Vietnam">Vietnam</option>
+                        <option value="Virgin Islands (Brit)">
+                          Virgin Islands (Brit)
+                        </option>
+                        <option value="Virgin Islands (USA)">
+                          Virgin Islands (USA)
+                        </option>
+                        <option value="Wake Island">Wake Island</option>
+                        <option value="Wallis &amp; Futana Is">
+                          Wallis &amp; Futana Is
+                        </option>
+                        <option value="Yemen">Yemen</option>
+                        <option value="Zaire">Zaire</option>
+                        <option value="Zambia">Zambia</option>
+                        <option value="Zimbabwe">Zimbabwe</option>
+                      </Select>
+                    </FormControl>
+                  </SimpleGrid>
+
+                  <FormControl>
+                    <FormLabel
+                      fontSize="sm"
+                      fontWeight="md"
+                      color={useColorModeValue('gray.700', 'gray.50')}
+                    >
+                      Photo / Logo
+                    </FormLabel>
+                    <SimpleGrid style={imgUploadContainer as CSSProperties}>
+                      <aside style={thumbsContainer as CSSProperties}>
+                        {files.map((f, i) => (
+                          <Box key={`${i}.${f.name}`}>
+                            <Avatar size="md" ml={5} src={f.preview} />
+                            <Text style={thumbLabel as CSSProperties}>
+                              {f.name} -{' '}
+                              {Math.floor(Math.log(f.size) / Math.log(1024))}
+                              mb
+                            </Text>
+                          </Box>
+                        ))}
+                      </aside>
+
+                      <SimpleGrid style={dropzoneContainer}>
+                        <Box
+                          {...getRootProps({
+                            className: 'dropzone',
+                          })}
+                          style={dropzone}
+                        >
+                          <input {...getInputProps()} />
+                          <Text
+                            color={useColorModeValue('gray.500', 'gray.400')}
+                          >
+                            Drag 'n' drop some files here, or click to select
+                            files
+                          </Text>
+                        </Box>
+                      </SimpleGrid>
+                    </SimpleGrid>
+                  </FormControl>
+
+                  <FormControl>
+                    <FormLabel
+                      fontSize="sm"
+                      fontWeight="md"
+                      color={useColorModeValue('gray.700', 'gray.50')}
+                    >
+                      Record Campaign Video
+                    </FormLabel>
+                    <Box>
+                      <LoomRecordButton
+                        onVideoChange={(url) => setVideoUrl(url)}
+                      />
+                    </Box>
+                  </FormControl>
+                </Stack>
             </chakra.form>
-          </GridItem>
-        </SimpleGrid>
-      </Box>
-      <Box visibility={{ base: "hidden", sm: "visible" }} aria-hidden="true">
-        <Box py={5}>
-          <Box
-            borderTop="solid 1px"
-            borderTopColor={useColorModeValue("gray.200", "whiteAlpha.200")}
-          ></Box>
-        </Box>
-      </Box>
-
-      <Box mt={[10, 0]}>
-        <SimpleGrid
-          display={{ base: "initial", md: "grid" }}
-          columns={{ md: 3 }}
-          spacing={{ md: 6 }}
-        >
-          <GridItem colSpan={{ md: 1 }}>
-            <Box px={[4, 0]}>
-              <Heading fontSize="lg" fontWeight="medium" lineHeight="6">
-                Personal Information
-              </Heading>
-              <Text
-                mt={1}
-                fontSize="sm"
-                color={useColorModeValue("gray.600", "gray.400")}
-              >
-                Use a permanent address where you can receive mail.
-              </Text>
-            </Box>
-          </GridItem>
-          <GridItem mt={[5, null, 0]} colSpan={{ md: 2 }}>
-            <chakra.form
-             
-              method="POST"
-              shadow="base"
-              rounded={[null, "md"]}
-              overflow={{ sm: "hidden" }}
-            >
-              <Stack
-                px={4}
-                py={5}
-                p={[null, 6]}
-                bg={useColorModeValue("white", "gray.700")}
-                spacing={6}
-              >
-                <SimpleGrid columns={6} spacing={6}>
-                  <FormControl as={GridItem} colSpan={[6, 3]}>
-                    <FormLabel
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      First name
-                    </FormLabel>
-                    <Input
-                      type="text"
-                      name="first_name"
-                      id="first_name"
-                      autoComplete="given-name"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    />
-                  </FormControl>
-
-                  <FormControl as={GridItem} colSpan={[6, 3]}>
-                    <FormLabel
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      Last name
-                    </FormLabel>
-                    <Input
-                      type="text"
-                      name="last_name"
-                      id="last_name"
-                      autoComplete="family-name"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    />
-                  </FormControl>
-
-                  <FormControl as={GridItem} colSpan={[6, 4]}>
-                    <FormLabel
-                      htmlFor="email_address"
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      Email address
-                    </FormLabel>
-                    <Input
-                      type="text"
-                      name="email_address"
-                      id="email_address"
-                      autoComplete="email"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    />
-                  </FormControl>
-
-                  <FormControl as={GridItem} colSpan={[6, 3]}>
-                    <FormLabel
-                      htmlFor="country"
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      Country / Region
-                    </FormLabel>
-                    <Select
-                      id="country"
-                      name="country"
-                      autoComplete="country"
-                      placeholder="Select option"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    >
-                      <option>United States</option>
-                      <option>Canada</option>
-                      <option>Mexico</option>
-                    </Select>
-                  </FormControl>
-
-                  <FormControl as={GridItem} colSpan={6}>
-                    <FormLabel
-                      htmlFor="street_address"
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      Street address
-                    </FormLabel>
-                    <Input
-                      type="text"
-                      name="street_address"
-                      id="street_address"
-                      autoComplete="street-address"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    />
-                  </FormControl>
-
-                  <FormControl as={GridItem} colSpan={[6, 6, null, 2]}>
-                    <FormLabel
-                      htmlFor="city"
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      City
-                    </FormLabel>
-                    <Input
-                      type="text"
-                      name="city"
-                      id="city"
-                      autoComplete="city"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    />
-                  </FormControl>
-
-                  <FormControl as={GridItem} colSpan={[6, 3, null, 2]}>
-                    <FormLabel
-                      htmlFor="state"
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      State / Province
-                    </FormLabel>
-                    <Input
-                      type="text"
-                      name="state"
-                      id="state"
-                      autoComplete="state"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    />
-                  </FormControl>
-
-                  <FormControl as={GridItem} colSpan={[6, 3, null, 2]}>
-                    <FormLabel
-                      htmlFor="postal_code"
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue("gray.700", "gray.50")}
-                    >
-                      ZIP / Postal
-                    </FormLabel>
-                    <Input
-                      type="text"
-                      name="postal_code"
-                      id="postal_code"
-                      autoComplete="postal-code"
-                      mt={1}
-                      focusBorderColor="brand.400"
-                      shadow="sm"
-                      size="sm"
-                      w="full"
-                      rounded="md"
-                    />
-                  </FormControl>
-                </SimpleGrid>
-              </Stack>
-              <Box
-                px={{ base: 4, sm: 6 }}
-                py={3}
-                bg={useColorModeValue("gray.50", "gray.900")}
-                textAlign="right"
-              >
-                <Button
-                  type="submit"
-                  colorScheme="brand"
-                  _focus={{ shadow: "" }}
-                  fontWeight="md"
-                >
-                  Save
-                </Button>
-              </Box>
-            </chakra.form>
-          </GridItem>
+        </GridItem>
         </SimpleGrid>
       </Box>
     </Box>
