+++ src/pages/createcampaign.tsx
@@ -1,4 +1,4 @@
-import React, { CSSProperties, Suspense, useEffect, useState } from 'react'
+import React, { Suspense, useEffect, useState } from 'react'
 import {
   chakra,
   Box,
@@ -29,7 +29,7 @@ import {
 } from '@chakra-ui/react'
 import { CheckIcon, CloseIcon, EditIcon, EmailIcon } from '@chakra-ui/icons'
 import { useEthers } from '@usedapp/core'
-import { useDropzone } from 'react-dropzone'
+
 import { useFormik } from 'formik'
 import { ContractFactory, ethers, providers } from 'ethers'
 import {
@@ -38,76 +38,16 @@ import {
   PoolMetadata,
 } from '../services/textile/types'
 import { TextileInstance } from '../services/textile/textile'
-import LoomRecordButton from 'src/components/LoomRecordButton'
 import StepsVerticalComponent from '../components/stepper/StepsVerticalComponent'
 import CampaignForm from '../components/forms/CampaignForm'
 import PoolForm from '../components/forms/PoolForm'
 import NotificationForm from '../components/forms/NotificationForm'
 
-const abi = require('../contracts/Pool.abi')
-const bytecode = require('../contracts/Pool.bytecode')
-
-const imgUploadContainer: CSSProperties = {
-  marginLeft: '0.25rem',
-}
-
-const dropzoneContainer: CSSProperties = {
-  width: '450px',
-  height: '107px',
-  display: 'flex',
-  justifyItems: 'start',
-  alignItems: 'start',
-  backgroundImage: `url("data:image/svg+xml,%3csvg width='100%25' height='100%25' xmlns='http://www.w3.org/2000/svg'%3e%3crect width='100%25' height='100%25' fill='none' stroke='grey' stroke-width='3' stroke-dasharray='2%2c 6' stroke-dashoffset='0' stroke-linecap='square'/%3e%3c/svg%3e")`,
-}
-
-const dropzone: CSSProperties = {
-  marginTop: '2px',
-  padding: 'auto',
-  display: 'flex',
-  justifyContent: 'start',
-  alignContent: 'start',
-  alignItems: 'center',
-  justifyItems: 'center',
-  width: '446px',
-  height: '103px',
-  marginLeft: '2px',
-  paddingLeft: '1rem',
-}
-
-const thumbsContainer: CSSProperties = {
-  width: '450px',
-  marginBottom: '0.5rem',
-  display: 'flex',
-  alignItems: 'center',
-  flexDirection: 'row',
-  flexWrap: 'wrap',
-}
-
-const thumbLabel: CSSProperties = {
-  marginLeft: '0.75rem',
-  marginRight: '1.25rem',
-  display: 'flex',
-  alignItems: 'center',
-  justifyItems: 'center',
-}
-
 export default function Component() {
-  const [videoUrl, setVideoUrl] = useState('')
-  const [poolDeployable, setPoolDeployable] = useState(false)
 
   const { account, library } = useEthers()
 
-  const { getRootProps, getInputProps } = useDropzone({
-    onDrop: (acceptedFiles) => {
-      setFiles(
-        acceptedFiles.map((file) =>
-          Object.assign(file, {
-            preview: URL.createObjectURL(file),
-          })
-        )
-      )
-    },
-  })
+  
 
   const campaignForm = useFormik({
     initialValues: {
@@ -150,7 +90,7 @@ export default function Component() {
       submissionLength: 432000000,
     },
     onSubmit: (values: PoolMetadata) => {
-      handleSavePoolDetails(values)
+      // handleSavePoolDetails(values)
     },
   })
 
@@ -164,7 +104,7 @@ export default function Component() {
       push: '0',
     },
     onSubmit: (values: CampaignSettings) => {
-      handleSaveNotificationPreferences(values)
+      // handleSaveNotificationPreferences(values)
     },
   })
 
@@ -190,7 +130,7 @@ export default function Component() {
             userCampaign.activePool
           )
           poolForm.setValues(activePool)
-          setPoolDeployable(true)
+          // setPoolDeployable(true) // PoolForm.tsx
         }
         console.log({
           campaign: campaignForm.values,
@@ -222,73 +162,73 @@ export default function Component() {
     campaignForm.setValues(campaignMetadata)
   }
 
-  const handleSavePoolDetails = async (values: PoolMetadata) => {
-    const textileInstance = await TextileInstance.getInstance()
+  // const handleSavePoolDetails = async (values: PoolMetadata) => {
+  //   const textileInstance = await TextileInstance.getInstance()
 
-    const pool = {
-      ...values,
-    }
+  //   const pool = {
+  //     ...values,
+  //   }
 
-    const poolMetadata = await textileInstance.uploadPoolMetadata({
-      poolAddress: poolForm.values.poolAddress
-        ? poolForm.values.poolAddress
-        : '',
-      ...pool,
-    })
+  //   const poolMetadata = await textileInstance.uploadPoolMetadata({
+  //     poolAddress: poolForm.values.poolAddress
+  //       ? poolForm.values.poolAddress
+  //       : '',
+  //     ...pool,
+  //   })
 
-    console.log({ poolMetadata })
+  //   console.log({ poolMetadata })
 
-    await textileInstance.uploadAndSetCampaignPool(
-      campaignForm.values._id,
-      poolMetadata
-    )
+  //   await textileInstance.uploadAndSetCampaignPool(
+  //     campaignForm.values._id,
+  //     poolMetadata
+  //   )
 
-    poolForm.setValues(poolMetadata)
-  }
+  //   poolForm.setValues(poolMetadata)
+  // } POOLFORM.TSX
 
-  const handleDeploy = async () => {
-    const signer = await (library as providers.Web3Provider).getSigner()
+  // const handleDeploy = async () => {
+  //   const signer = await (library as providers.Web3Provider).getSigner()
 
-    const pool = {
-      ...poolForm.values,
-    }
+  //   const pool = {
+  //     ...poolForm.values,
+  //   }
 
-    const poolFactory = new ContractFactory(abi, bytecode, signer)
+  //   const poolFactory = new ContractFactory(abi, bytecode, signer)
 
-    console.log({ pool, poolFactory })
+  //   console.log({ pool, poolFactory })
 
-    const contract = await poolFactory.deploy(
-      pool.poolName,
-      campaignForm.values.name,
-      ethers.BigNumber.from(pool.capital),
-      ethers.utils.getAddress(pool.capitalAddress),
-      ethers.utils.getAddress(pool.nftAddress),
-      ethers.utils.getAddress(pool.poolOwner),
-      ethers.utils.getAddress(pool.rngAddress),
-      ethers.BigNumber.from(pool.campaignLength),
-      ethers.BigNumber.from(pool.votingLength),
-      ethers.BigNumber.from(pool.decisionLength),
-      ethers.BigNumber.from(pool.submissionLength)
-    )
+  //   const contract = await poolFactory.deploy(
+  //     pool.poolName,
+  //     campaignForm.values.name,
+  //     ethers.BigNumber.from(pool.capital),
+  //     ethers.utils.getAddress(pool.capitalAddress),
+  //     ethers.utils.getAddress(pool.nftAddress),
+  //     ethers.utils.getAddress(pool.poolOwner),
+  //     ethers.utils.getAddress(pool.rngAddress),
+  //     ethers.BigNumber.from(pool.campaignLength),
+  //     ethers.BigNumber.from(pool.votingLength),
+  //     ethers.BigNumber.from(pool.decisionLength),
+  //     ethers.BigNumber.from(pool.submissionLength)
+  //   )
 
-    poolForm.values.poolAddress = contract.address
-  }
+  //   poolForm.values.poolAddress = contract.address
+  // } POOLFORM.TSX
 
-  const handleSaveNotificationPreferences = async (
-    values: CampaignSettings
-  ) => {
-    const textileInstance = await TextileInstance.getInstance()
+  // const handleSaveNotificationPreferences = async (
+  //   values: CampaignSettings
+  // ) => {
+  //   const textileInstance = await TextileInstance.getInstance()
 
-    // const preferenceMetadata =
-    //     await textileInstance.uploadCampaignPreferences({
-    //         ...values,
-    //     });
+  //   // const preferenceMetadata =
+  //   //     await textileInstance.uploadCampaignPreferences({
+  //   //         ...values,
+  //   //     });
 
-    await textileInstance.setCampaignPreferences(
-      campaignForm.values._id,
-      values
-    )
-  }
+  //   await textileInstance.setCampaignPreferences(
+  //     campaignForm.values._id,
+  //     values
+  //   )
+  // } // NOTIFICATIONFORM.TSX 
 
   // function EditableControls() {
   //   const {
@@ -371,13 +311,6 @@ export default function Component() {
                 // overflow={{ sm: "hidden" }}
                 onSubmit={campaignForm.handleSubmit}
               >
-                {/* <Stack
-                  px={4}
-                  py={5}
-                  bg={useColorModeValue('white', 'gray.700')}
-                  spacing={6}
-                  p={{ sm: 6 }}
-                > */}
                 <StepsVerticalComponent
                   steps={[
                     { label: 'Campaign 📜' },
@@ -388,560 +321,13 @@ export default function Component() {
                   <CampaignForm />
                   <PoolForm />
                   <NotificationForm />
-                  {/* <SimpleGrid columns={3} spacing={6}>
-                    <FormControl as={GridItem} colSpan={[3, 2]}>
-                      <FormLabel
-                        fontSize="sm"
-                        fontWeight="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                        htmlFor="brandName"
-                      >
-                        Brand Name
-                      </FormLabel>
-                      <Input
-                        value={campaignForm.values.brandName}
-                        onChange={campaignForm.handleChange}
-                        type="text"
-                        placeholder="Company Name"
-                        id="brandName"
-                        name="brandName"
-                        focusBorderColor="brand.400"
-                        rounded="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      />
-                    </FormControl>
-                  </SimpleGrid>
-                  <SimpleGrid columns={3} spacing={6}>
-                    <FormControl as={GridItem} colSpan={[3, 2]}>
-                      <FormLabel
-                        fontSize="sm"
-                        fontWeight="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      >
-                        Brand Website
-                      </FormLabel>
-                      <InputGroup size="sm">
-                        <InputLeftAddon
-                          children="https://"
-                          bg={useColorModeValue('gray.50', 'gray.800')}
-                          color={useColorModeValue('gray.500', 'gay.50')}
-                          rounded="md"
-                        />
-                        <Input
-                          value={campaignForm.values.brandWebsite}
-                          onChange={campaignForm.handleChange}
-                          type="text"
-                          placeholder="www.example.com"
-                          id="brandWebsite"
-                          name="brandWebsite"
-                          focusBorderColor="brand.400"
-                          rounded="md"
-                          color={useColorModeValue('gray.700', 'gray.50')}
-                        />
-                      </InputGroup>
-                    </FormControl>
-                  </SimpleGrid>
-
-                  <SimpleGrid columns={3} spacing={6}>
-                    <FormControl as={GridItem} colSpan={[3, 2]}>
-                      <FormLabel
-                        fontSize="sm"
-                        fontWeight="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      >
-                        Twitter Username
-                      </FormLabel>
-                      <InputGroup size="sm">
-                        <InputLeftAddon
-                          children="@"
-                          bg={useColorModeValue('gray.50', 'gray.800')}
-                          color={useColorModeValue('gray.500', 'gay.50')}
-                          rounded="md"
-                        />
-                        <Input
-                          value={campaignForm.values.twitterAccount}
-                          onChange={campaignForm.handleChange}
-                          id="twitterAccount"
-                          name="twitterAccount"
-                          placeholder="creativecrtv"
-                          focusBorderColor="brand.400"
-                          rounded="md"
-                          color={useColorModeValue('gray.700', 'gray.50')}
-                        />
-                      </InputGroup>
-                      <FormHelperText>
-                        Verify your account via Twitter (0.1 LINK).
-                      </FormHelperText>
-                    </FormControl>
-                  </SimpleGrid>
-
-                  <SimpleGrid columns={3} spacing={6}>
-                    <FormControl mt={1} as={GridItem} colSpan={[3, 2]}>
-                      <FormLabel
-                        htmlFor="campaignBrief"
-                        fontSize="sm"
-                        fontWeight="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      >
-                        Campaign Brief
-                      </FormLabel>
-                      <Textarea
-                        value={campaignForm.values.campaignBrief}
-                        onChange={campaignForm.handleChange}
-                        id="campaignBrief"
-                        name="campaignBrief"
-                        placeholder="Additional information about your campaign"
-                        mt={1}
-                        rows={3}
-                        shadow="sm"
-                        focusBorderColor="brand.400"
-                        fontSize={{
-                          sm: 'sm',
-                        }}
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      />
-                      <FormHelperText>
-                        Brief description for your requirements. URLs are
-                        hyperlinked.
-                      </FormHelperText>
-                    </FormControl>
-                  </SimpleGrid>
-
-                  <SimpleGrid columns={3} spacing={6}>
-                    <FormControl as={GridItem} colSpan={[3, 2]}>
-                      <FormLabel
-                        htmlFor="email"
-                        fontSize="sm"
-                        fontWeight="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      >
-                        Email address
-                      </FormLabel>
-                      <InputGroup size={'sm'}>
-                        <InputLeftAddon
-                          children={<EmailIcon />}
-                          bg={useColorModeValue('gray.50', 'gray.800')}
-                          color={useColorModeValue('gray.500', 'gay.50')}
-                          rounded="md"
-                        />
-                        <Input
-                          type="text"
-                          name="email"
-                          id="email"
-                          value={campaignForm.values.email}
-                          onChange={campaignForm.handleChange}
-                          autoComplete="email"
-                          focusBorderColor="brand.400"
-                          w="full"
-                          rounded="md"
-                          color={useColorModeValue('gray.700', 'gray.50')}
-                        />
-                      </InputGroup>
-                    </FormControl>
-                  </SimpleGrid>
-
-                  <SimpleGrid columns={3} spacing={6}>
-                    <FormControl>
-                      <FormLabel
-                        htmlFor="country"
-                        fontSize="sm"
-                        fontWeight="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      >
-                        Country / Region
-                      </FormLabel>
-                      <Select
-                        value={campaignForm.values.country}
-                        onChange={campaignForm.handleChange}
-                        id="country"
-                        name="country"
-                        autoComplete="country"
-                        placeholder="select country"
-                        mt={1}
-                        focusBorderColor="brand.400"
-                        shadow="sm"
-                        size="sm"
-                        w="full"
-                        rounded="md"
-                        color={useColorModeValue('gray.700', 'gray.50')}
-                      >
-                        <option value="Afganistan">Afghanistan</option>
-                        <option value="Albania">Albania</option>
-                        <option value="Algeria">Algeria</option>
-                        <option value="American Samoa">American Samoa</option>
-                        <option value="Andorra">Andorra</option>
-                        <option value="Angola">Angola</option>
-                        <option value="Anguilla">Anguilla</option>
-                        <option value="Antigua &amp; Barbuda">
-                          Antigua &amp; Barbuda
-                        </option>
-                        <option value="Argentina">Argentina</option>
-                        <option value="Armenia">Armenia</option>
-                        <option value="Aruba">Aruba</option>
-                        <option value="Australia">Australia</option>
-                        <option value="Austria">Austria</option>
-                        <option value="Azerbaijan">Azerbaijan</option>
-                        <option value="Bahamas">Bahamas</option>
-                        <option value="Bahrain">Bahrain</option>
-                        <option value="Bangladesh">Bangladesh</option>
-                        <option value="Barbados">Barbados</option>
-                        <option value="Belarus">Belarus</option>
-                        <option value="Belgium">Belgium</option>
-                        <option value="Belize">Belize</option>
-                        <option value="Benin">Benin</option>
-                        <option value="Bermuda">Bermuda</option>
-                        <option value="Bhutan">Bhutan</option>
-                        <option value="Bolivia">Bolivia</option>
-                        <option value="Bonaire">Bonaire</option>
-                        <option value="Bosnia &amp; Herzegovina">
-                          Bosnia &amp; Herzegovina
-                        </option>
-                        <option value="Botswana">Botswana</option>
-                        <option value="Brazil">Brazil</option>
-                        <option value="British Indian Ocean Ter">
-                          British Indian Ocean Ter
-                        </option>
-                        <option value="Brunei">Brunei</option>
-                        <option value="Bulgaria">Bulgaria</option>
-                        <option value="Burkina Faso">Burkina Faso</option>
-                        <option value="Burundi">Burundi</option>
-                        <option value="Cambodia">Cambodia</option>
-                        <option value="Cameroon">Cameroon</option>
-                        <option value="Canada">Canada</option>
-                        <option value="Canary Islands">Canary Islands</option>
-                        <option value="Cape Verde">Cape Verde</option>
-                        <option value="Cayman Islands">Cayman Islands</option>
-                        <option value="Central African Republic">
-                          Central African Republic
-                        </option>
-                        <option value="Chad">Chad</option>
-                        <option value="Channel Islands">Channel Islands</option>
-                        <option value="Chile">Chile</option>
-                        <option value="China">China</option>
-                        <option value="Christmas Island">
-                          Christmas Island
-                        </option>
-                        <option value="Cocos Island">Cocos Island</option>
-                        <option value="Colombia">Colombia</option>
-                        <option value="Comoros">Comoros</option>
-                        <option value="Congo">Congo</option>
-                        <option value="Cook Islands">Cook Islands</option>
-                        <option value="Costa Rica">Costa Rica</option>
-                        <option value="Cote DIvoire">Cote DIvoire</option>
-                        <option value="Croatia">Croatia</option>
-                        <option value="Cuba">Cuba</option>
-                        <option value="Curaco">Curacao</option>
-                        <option value="Cyprus">Cyprus</option>
-                        <option value="Czech Republic">Czech Republic</option>
-                        <option value="Denmark">Denmark</option>
-                        <option value="Djibouti">Djibouti</option>
-                        <option value="Dominica">Dominica</option>
-                        <option value="Dominican Republic">
-                          Dominican Republic
-                        </option>
-                        <option value="East Timor">East Timor</option>
-                        <option value="Ecuador">Ecuador</option>
-                        <option value="Egypt">Egypt</option>
-                        <option value="El Salvador">El Salvador</option>
-                        <option value="Equatorial Guinea">
-                          Equatorial Guinea
-                        </option>
-                        <option value="Eritrea">Eritrea</option>
-                        <option value="Estonia">Estonia</option>
-                        <option value="Ethiopia">Ethiopia</option>
-                        <option value="Falkland Islands">
-                          Falkland Islands
-                        </option>
-                        <option value="Faroe Islands">Faroe Islands</option>
-                        <option value="Fiji">Fiji</option>
-                        <option value="Finland">Finland</option>
-                        <option value="France">France</option>
-                        <option value="French Guiana">French Guiana</option>
-                        <option value="French Polynesia">
-                          French Polynesia
-                        </option>
-                        <option value="French Southern Ter">
-                          French Southern Ter
-                        </option>
-                        <option value="Gabon">Gabon</option>
-                        <option value="Gambia">Gambia</option>
-                        <option value="Georgia">Georgia</option>
-                        <option value="Germany">Germany</option>
-                        <option value="Ghana">Ghana</option>
-                        <option value="Gibraltar">Gibraltar</option>
-                        <option value="Great Britain">Great Britain</option>
-                        <option value="Greece">Greece</option>
-                        <option value="Greenland">Greenland</option>
-                        <option value="Grenada">Grenada</option>
-                        <option value="Guadeloupe">Guadeloupe</option>
-                        <option value="Guam">Guam</option>
-                        <option value="Guatemala">Guatemala</option>
-                        <option value="Guinea">Guinea</option>
-                        <option value="Guyana">Guyana</option>
-                        <option value="Haiti">Haiti</option>
-                        <option value="Hawaii">Hawaii</option>
-                        <option value="Honduras">Honduras</option>
-                        <option value="Hong Kong">Hong Kong</option>
-                        <option value="Hungary">Hungary</option>
-                        <option value="Iceland">Iceland</option>
-                        <option value="Indonesia">Indonesia</option>
-                        <option value="India">India</option>
-                        <option value="Iran">Iran</option>
-                        <option value="Iraq">Iraq</option>
-                        <option value="Ireland">Ireland</option>
-                        <option value="Isle of Man">Isle of Man</option>
-                        <option value="Israel">Israel</option>
-                        <option value="Italy">Italy</option>
-                        <option value="Jamaica">Jamaica</option>
-                        <option value="Japan">Japan</option>
-                        <option value="Jordan">Jordan</option>
-                        <option value="Kazakhstan">Kazakhstan</option>
-                        <option value="Kenya">Kenya</option>
-                        <option value="Kiribati">Kiribati</option>
-                        <option value="Korea North">Korea North</option>
-                        <option value="Korea Sout">Korea South</option>
-                        <option value="Kuwait">Kuwait</option>
-                        <option value="Kyrgyzstan">Kyrgyzstan</option>
-                        <option value="Laos">Laos</option>
-                        <option value="Latvia">Latvia</option>
-                        <option value="Lebanon">Lebanon</option>
-                        <option value="Lesotho">Lesotho</option>
-                        <option value="Liberia">Liberia</option>
-                        <option value="Libya">Libya</option>
-                        <option value="Liechtenstein">Liechtenstein</option>
-                        <option value="Lithuania">Lithuania</option>
-                        <option value="Luxembourg">Luxembourg</option>
-                        <option value="Macau">Macau</option>
-                        <option value="Macedonia">Macedonia</option>
-                        <option value="Madagascar">Madagascar</option>
-                        <option value="Malaysia">Malaysia</option>
-                        <option value="Malawi">Malawi</option>
-                        <option value="Maldives">Maldives</option>
-                        <option value="Mali">Mali</option>
-                        <option value="Malta">Malta</option>
-                        <option value="Marshall Islands">
-                          Marshall Islands
-                        </option>
-                        <option value="Martinique">Martinique</option>
-                        <option value="Mauritania">Mauritania</option>
-                        <option value="Mauritius">Mauritius</option>
-                        <option value="Mayotte">Mayotte</option>
-                        <option value="Mexico">Mexico</option>
-                        <option value="Midway Islands">Midway Islands</option>
-                        <option value="Moldova">Moldova</option>
-                        <option value="Monaco">Monaco</option>
-                        <option value="Mongolia">Mongolia</option>
-                        <option value="Montserrat">Montserrat</option>
-                        <option value="Morocco">Morocco</option>
-                        <option value="Mozambique">Mozambique</option>
-                        <option value="Myanmar">Myanmar</option>
-                        <option value="Nambia">Nambia</option>
-                        <option value="Nauru">Nauru</option>
-                        <option value="Nepal">Nepal</option>
-                        <option value="Netherland Antilles">
-                          Netherland Antilles
-                        </option>
-                        <option value="Netherlands">
-                          Netherlands (Holland, Europe)
-                        </option>
-                        <option value="Nevis">Nevis</option>
-                        <option value="New Caledonia">New Caledonia</option>
-                        <option value="New Zealand">New Zealand</option>
-                        <option value="Nicaragua">Nicaragua</option>
-                        <option value="Niger">Niger</option>
-                        <option value="Nigeria">Nigeria</option>
-                        <option value="Niue">Niue</option>
-                        <option value="Norfolk Island">Norfolk Island</option>
-                        <option value="Norway">Norway</option>
-                        <option value="Oman">Oman</option>
-                        <option value="Pakistan">Pakistan</option>
-                        <option value="Palau Island">Palau Island</option>
-                        <option value="Palestine">Palestine</option>
-                        <option value="Panama">Panama</option>
-                        <option value="Papua New Guinea">
-                          Papua New Guinea
-                        </option>
-                        <option value="Paraguay">Paraguay</option>
-                        <option value="Peru">Peru</option>
-                        <option value="Phillipines">Philippines</option>
-                        <option value="Pitcairn Island">Pitcairn Island</option>
-                        <option value="Poland">Poland</option>
-                        <option value="Portugal">Portugal</option>
-                        <option value="Puerto Rico">Puerto Rico</option>
-                        <option value="Qatar">Qatar</option>
-                        <option value="Republic of Montenegro">
-                          Republic of Montenegro
-                        </option>
-                        <option value="Republic of Serbia">
-                          Republic of Serbia
-                        </option>
-                        <option value="Reunion">Reunion</option>
-                        <option value="Romania">Romania</option>
-                        <option value="Russia">Russia</option>
-                        <option value="Rwanda">Rwanda</option>
-                        <option value="St Barthelemy">St Barthelemy</option>
-                        <option value="St Eustatius">St Eustatius</option>
-                        <option value="St Helena">St Helena</option>
-                        <option value="St Kitts-Nevis">St Kitts-Nevis</option>
-                        <option value="St Lucia">St Lucia</option>
-                        <option value="St Maarten">St Maarten</option>
-                        <option value="St Pierre &amp; Miquelon">
-                          St Pierre &amp; Miquelon
-                        </option>
-                        <option value="St Vincent &amp; Grenadines">
-                          St Vincent &amp; Grenadines
-                        </option>
-                        <option value="Saipan">Saipan</option>
-                        <option value="Samoa">Samoa</option>
-                        <option value="Samoa American">Samoa American</option>
-                        <option value="San Marino">San Marino</option>
-                        <option value="Sao Tome &amp; Principe">
-                          Sao Tome &amp; Principe
-                        </option>
-                        <option value="Saudi Arabia">Saudi Arabia</option>
-                        <option value="Senegal">Senegal</option>
-                        <option value="Seychelles">Seychelles</option>
-                        <option value="Sierra Leone">Sierra Leone</option>
-                        <option value="Singapore">Singapore</option>
-                        <option value="Slovakia">Slovakia</option>
-                        <option value="Slovenia">Slovenia</option>
-                        <option value="Solomon Islands">Solomon Islands</option>
-                        <option value="Somalia">Somalia</option>
-                        <option value="South Africa">South Africa</option>
-                        <option value="Spain">Spain</option>
-                        <option value="Sri Lanka">Sri Lanka</option>
-                        <option value="Sudan">Sudan</option>
-                        <option value="Suriname">Suriname</option>
-                        <option value="Swaziland">Swaziland</option>
-                        <option value="Sweden">Sweden</option>
-                        <option value="Switzerland">Switzerland</option>
-                        <option value="Syria">Syria</option>
-                        <option value="Tahiti">Tahiti</option>
-                        <option value="Taiwan">Taiwan</option>
-                        <option value="Tajikistan">Tajikistan</option>
-                        <option value="Tanzania">Tanzania</option>
-                        <option value="Thailand">Thailand</option>
-                        <option value="Togo">Togo</option>
-                        <option value="Tokelau">Tokelau</option>
-                        <option value="Tonga">Tonga</option>
-                        <option value="Trinidad &amp; Tobago">
-                          Trinidad &amp; Tobago
-                        </option>
-                        <option value="Tunisia">Tunisia</option>
-                        <option value="Turkey">Turkey</option>
-                        <option value="Turkmenistan">Turkmenistan</option>
-                        <option value="Turks &amp; Caicos Is">
-                          Turks &amp; Caicos Is
-                        </option>
-                        <option value="Tuvalu">Tuvalu</option>
-                        <option value="Uganda">Uganda</option>
-                        <option value="United Kingdom">United Kingdom</option>
-                        <option value="Ukraine">Ukraine</option>
-                        <option value="United Arab Erimates">
-                          United Arab Emirates
-                        </option>
-                        <option value="United States of America">
-                          United States of America
-                        </option>
-                        <option value="Uraguay">Uruguay</option>
-                        <option value="Uzbekistan">Uzbekistan</option>
-                        <option value="Vanuatu">Vanuatu</option>
-                        <option value="Vatican City State">
-                          Vatican City State
-                        </option>
-                        <option value="Venezuela">Venezuela</option>
-                        <option value="Vietnam">Vietnam</option>
-                        <option value="Virgin Islands (Brit)">
-                          Virgin Islands (Brit)
-                        </option>
-                        <option value="Virgin Islands (USA)">
-                          Virgin Islands (USA)
-                        </option>
-                        <option value="Wake Island">Wake Island</option>
-                        <option value="Wallis &amp; Futana Is">
-                          Wallis &amp; Futana Is
-                        </option>
-                        <option value="Yemen">Yemen</option>
-                        <option value="Zaire">Zaire</option>
-                        <option value="Zambia">Zambia</option>
-                        <option value="Zimbabwe">Zimbabwe</option>
-                      </Select>
-                    </FormControl>
-                  </SimpleGrid>
-
-                  <FormControl>
-                    <FormLabel
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue('gray.700', 'gray.50')}
-                    >
-                      Photo / Logo
-                    </FormLabel>
-                    <SimpleGrid style={imgUploadContainer as CSSProperties}>
-                      <aside style={thumbsContainer as CSSProperties}>
-                        {files.map((f, i) => (
-                          <Box key={`${i}.${f.name}`}>
-                            <Avatar size="md" ml={5} src={f.preview} />
-                            <Text style={thumbLabel as CSSProperties}>
-                              {f.name} -{' '}
-                              {Math.floor(Math.log(f.size) / Math.log(1024))}
-                              mb
-                            </Text>
-                          </Box>
-                        ))}
-                      </aside>
-
-                      <SimpleGrid style={dropzoneContainer}>
-                        <Box
-                          {...getRootProps({
-                            className: 'dropzone',
-                          })}
-                          style={dropzone}
-                        >
-                          <input {...getInputProps()} />
-                          <Text
-                            color={useColorModeValue('gray.500', 'gray.400')}
-                          >
-                            Drag 'n' drop some files here, or click to select
-                            files
-                          </Text>
-                        </Box>
-                      </SimpleGrid>
-                    </SimpleGrid>
-                  </FormControl>
-
-                  <FormControl>
-                    <FormLabel
-                      fontSize="sm"
-                      fontWeight="md"
-                      color={useColorModeValue('gray.700', 'gray.50')}
-                    >
-                      Record Campaign Video
-                    </FormLabel>
-                    <Box>
-                      <LoomRecordButton
-                        onVideoChange={(url) => setVideoUrl(url)}
-                      />
-                    </Box>
-                  </FormControl> */}
-                  </StepsVerticalComponent>
-                {/* </Stack> */}
+                </StepsVerticalComponent>
                 <Box
                   px={{ base: 4, sm: 6 }}
                   py={3}
                   bg={useColorModeValue('gray.50', 'gray.900')}
                   textAlign="right"
                 >
-                  <Button
-                    type="submit"
-                    colorScheme="brand"
-                    _focus={{ shadow: '' }}
-                    fontWeight="md"
-                    color={useColorModeValue('gray.700', 'gray.50')}
-                    // onClick={handleSaveCampaignDetails}
-                  >
-                    Save
-                  </Button>
                 </Box>
               </form>
             </GridItem>
